Refactor OIDC tests
Add helper functions to the LDAP mock to create test users and
group memberships, and use those everywhere. Remove duplicate tests
and now-unused configuration files. Consolidate all of the OIDC
login tests into a single test file.
