--- tests/data/config/oidc-gid.yaml.in
@@ -1,29 +1,0 @@
-realm: "example.com"
-logLevel: "DEBUG"
-sessionSecretFile: "{session_secret_file}"
-databaseUrl: "{database_url}"
-redisUrl: "redis://localhost:6379/0"
-initialAdmins: ["admin"]
-afterLogoutUrl: "https://example.com/landing"
-groupMapping:
-  "exec:admin": ["admin"]
-  "exec:test": ["test"]
-  "read:all": ["foo", "admin", "org-a-team"]
-knownScopes:
-  "admin:token": "Can create and modify tokens for any user"
-  "exec:admin": "admin description"
-  "exec:test": "test description"
-  "read:all": "can read everything"
-  "user:token": "Can create and modify user tokens"
-ldap:
-  url: "ldaps://ldap.example.com/"
-  groupBaseDn: "dc=example,dc=com"
-  userBaseDn: "ou=people,dc=example,dc=com"
-oidc:
-  clientId: "some-oidc-client-id"
-  clientSecretFile: "{oidc_secret_file}"
-  loginUrl: "https://upstream.example.com/oidc/login"
-  redirectUrl: "https://upstream.example.com/login"
-  tokenUrl: "https://upstream.example.com/token"
-  issuer: "https://upstream.example.com/"
-  audience: "https://test.example.com/"
